scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
lxml==5.3.0
chromadb==0.5.23
